Washington D.C. Commuters 2020

Washington D.C. inbound and outbound commuters by county 2020

Usage

commuters

Format

A data frame with 97 rows and 10 columns

one

one

Source

https://www.census.gov/data/tables/2020/demo/metro-micro/commuting-flows-2020.html Data was created by the U.S. government and is not subject to copyright in the United States pursuant to 17 U.S.C.  105.

Details

From the U.S. Census Bureau: "The American Community Survey (ACS) asks respondents about their primary workplace location. When information about workers’ residence location and workplace location are coupled, a commuting flow is generated. The origin-destination flow format informs our understanding of interconnectedness between communities, including the interchange of people, goods, and services." For additional information, see the page on commuter data.

References

U.S. Census Bureau, 2016-2020 5-Year ACS Commuting Flows (2020) last accessed October 24, 2024.

Examples

commuters
#> # A tibble: 97 × 10
#>    from_state_fips_code from_county_fips_code from_state_name   from_county_name
#>    <chr>                <chr>                 <chr>             <chr>           
#>  1 06                   037                   California        Los Angeles Cou…
#>  2 06                   059                   California        Orange County   
#>  3 06                   073                   California        San Diego County
#>  4 10                   003                   Delaware          New Castle Coun…
#>  5 10                   005                   Delaware          Sussex County   
#>  6 11                   001                   District of Colu… District of Col…
#>  7 11                   001                   District of Colu… District of Col…
#>  8 11                   001                   District of Colu… District of Col…
#>  9 11                   001                   District of Colu… District of Col…
#> 10 11                   001                   District of Colu… District of Col…
#> # ℹ 87 more rows
#> # ℹ 6 more variables: to_state_fips_code <chr>, to_county_fips_code <chr>,
#> #   to_state_name <chr>, to_county_name <chr>, workers_in_commuting_flow <dbl>,
#> #   margin_of_error <dbl>