Washington D.C. commuter counties 2020

Washington D.C. inbound commuter counties 2020

Usage

dc_sites

Format

A data frame of 96 rows and four variables:

id

five character FIPS code for state and county

name

county name

lat

latitude

lon

longitude

Source

https://www.irs.gov/statistics/soi-tax-stats-migration-data-2021-2022

Details

From the U.S. Census Bureau: "The American Community Survey (ACS) asks respondents about their primary workplace location. When information about workers’ residence location and workplace location are coupled, a commuting flow is generated. The origin-destination flow format informs our understanding of interconnectedness between communities, including the interchange of people, goods, and services." For additional information, see the page on commuter data.

References

Internal Revenue Service, Statistics of Income Division (2022). SOI Tax Stats - Migration Data. Last accessed October 26, 2024.

Examples

dc_sites
#>       id                 name      lat        lon
#> 1  51683             Manassas 38.74960  -77.48478
#> 2  51600              Fairfax 38.85167  -77.30082
#> 3  51650              Hampton 37.03652  -76.35915
#> 4  51840           Winchester 39.17096  -78.17377
#> 5  51013            Arlington 38.87677  -77.10139
#> 6  51137               Orange 38.24532  -78.01505
#> 7  11001 District of Columbia 38.90417  -77.01654
#> 8  51043               Clarke 39.11294  -77.99185
#> 9  42001                Adams 39.87150  -77.21614
#> 10 24043           Washington 39.60765  -77.81777
#> 11 51810       Virginia Beach 36.73205  -76.04427
#> 12 24035         Queen Anne's 39.06194  -76.03564
#> 13 06059               Orange 33.70191 -117.76485
#> 14 51685        Manassas Park 38.77047  -77.44188
#> 15 51107              Loudoun 39.08984  -77.63475
#> 16 42041           Cumberland 40.16213  -77.26156
#> 17 51550           Chesapeake 36.67762  -76.30437
#> 18 24021            Frederick 39.47232  -77.39894
#> 19 12103             Pinellas 27.93109  -82.72284
#> 20 54027            Hampshire 39.32100  -78.60830
#> 21 25017            Middlesex 42.48685  -71.38800
#> 22 45051                Horry 33.92125  -78.99538
#> 23 51033             Caroline 38.03007  -77.34562
#> 24 24017              Charles 38.51091  -76.98581
#> 25 24011             Caroline 38.87155  -75.82907
#> 26 24025              Harford 39.54870  -76.30523
#> 27 42133                 York 39.92125  -76.73021
#> 28 24027               Howard 39.25324  -76.92547
#> 29 10003           New Castle 39.57964  -75.64065
#> 30 24009              Calvert 38.53946  -76.56803
#> 31 51179             Stafford 38.41850  -77.45897
#> 32 24510            Baltimore 39.30145  -76.61007
#> 33 24003         Anne Arundel 39.00424  -76.60110
#> 34 42029              Chester 39.97365  -75.74628
#> 35 12099           Palm Beach 26.64674  -80.46528
#> 36 51760             Richmond 37.52644  -77.47093
#> 37 24031           Montgomery 39.13709  -77.20262
#> 38 51041         Chesterfield 37.37491  -77.58923
#> 39 06073            San Diego 33.03437 -116.73854
#> 40 24013              Carroll 39.56607  -77.02398
#> 41 48113               Dallas 32.76668  -96.77784
#> 42 24033      Prince George's 38.83157  -76.84802
#> 43 51061             Fauquier 38.73293  -77.80702
#> 44 36047                Kings 40.63083  -73.94742
#> 45 51069            Frederick 39.20161  -78.25973
#> 46 42101         Philadelphia 40.00740  -75.13414
#> 47 34013                Essex 40.78668  -74.24528
#> 48 51510           Alexandria 38.81401  -77.08183
#> 49 36061             New York 40.77773  -73.96869
#> 50 51187               Warren 38.90864  -78.20265
#> 51 54037            Jefferson 39.30697  -77.86031
#> 52 42003            Allegheny 40.46890  -79.98443
#> 53 51059              Fairfax 38.83566  -77.27458
#> 54 51193         Westmoreland 38.11025  -76.80192
#> 55 37081             Guilford 36.07978  -79.78828
#> 56 51630       Fredericksburg 38.29893  -77.48679
#> 57 51085              Hanover 37.75925  -77.48880
#> 58 12011              Broward 26.15176  -80.48730
#> 59 42055             Franklin 39.92939  -77.72163
#> 60 51157         Rappahannock 38.68297  -78.15836
#> 61 54003             Berkeley 39.46705  -78.02366
#> 62 37183                 Wake 35.78800  -78.65091
#> 63 24005            Baltimore 39.45887  -76.62801
#> 64 51099          King George 38.26817  -77.15847
#> 65 51153       Prince William 38.70911  -77.48873
#> 66 37119          Mecklenburg 35.24395  -80.82809
#> 67 51610         Falls Church 38.88459  -77.17538
#> 68 51087              Henrico 37.53625  -77.40261
#> 69 51177         Spotsylvania 38.18588  -77.65509
#> 70 24029                 Kent 39.24976  -76.04762
#> 71 51003            Albemarle 38.02092  -78.55534
#> 72 06037          Los Angeles 34.30781 -118.22842
#> 73 24037           St. Mary's 38.30054  -76.60602
#> 74 24041               Talbot 38.76615  -76.10887
#> 75 51047             Culpeper 38.47890  -77.95941
#> 76 17031                 Cook 41.83998  -87.81672
#> 77 12095               Orange 28.51270  -81.32120
#> 78 10005               Sussex 38.66137  -75.39026
#> 79 37035              Catawba 35.66203  -81.21338