Global flight path and disappearance of Amelia Earhart 1937

Global flight path of Amelia Earhart 1937 and her disappearance in transit to Howland Island

Usage

earhart

Format

a data frame with 9 variables and 33 observations:

date

date

departure_city

departure_city

begin_lat

begin_lat

begin_lon

begin_lon

arrival_city

arrival_city

end_lat

end_lat

end_lon

end_lon

nautical_miles

nautical_miles

status

status

Source

https://en.wikipedia.org/wiki/Amelia_Earhart

Details

Amelia Earhart was an American aviation pioneer and author. Among her many aviation records, she was the first female pilot to fly solo across the Atlantic in 1928, to fly solo across the North American continent and back, and the first to fly from Hawaii to Oakland, California. In 1937, she and her navigator Fred Noonan planned to circumnavigate the globe in a Lockheed Model 10-E Electra. The itinerary included 33 stops. The trip was to begin in Oakland, California, and go eastward. After completing 30 of 33 stops, their plane went missing after they departed from Lae, New Guinea, in route to Howland Island, in the middle of the Pacific Ocean. The two were declared lost at sea. In 2024 an acoustic signature of an airplane was found within 100 miles of Howland Island at a depth of 16,000 ft (4.9km) of water. Some have speculated that the plane is Earhart's.

Examples

earhart
#> # A tibble: 33 × 9
#>    date       departure_city    begin_lat begin_lon arrival_city end_lat end_lon
#>    <date>     <chr>                 <dbl>     <dbl> <chr>          <dbl>   <dbl>
#>  1 1937-05-20 Oakland, Califor…     37.8     -122.  Burbank, Ca…   34.2   -118. 
#>  2 1937-05-21 Burbank, Califor…     34.2     -118.  Tucson, Ari…   32.2   -111. 
#>  3 1937-05-22 Tucson, Arizona       32.2     -111.  New Orleans…   30.0    -90.1
#>  4 1937-05-23 New Orleans, Lou…     30.0      -90.1 Miami, Flor…   25.8    -80.2
#>  5 1937-06-01 Miami, Florida        25.8      -80.2 San Juan, P…   18.5    -66.1
#>  6 1937-06-02 San Juan, Puerto…     18.5      -66.1 Caripito, V…   10.1    -63.1
#>  7 1937-06-03 Caripito, Venezu…     10.1      -63.1 Paramaribo,…    5.82   -55.2
#>  8 1937-06-04 Paramaribo, Suri…      5.82     -55.2 Fortaleza, …   -3.73   -38.5
#>  9 1937-06-05 Fortaleza, Brazil     -3.73     -38.5 Natal, Braz…   -5.81   -35.2
#> 10 1937-06-07 Natal, Brazil         -5.81     -35.2 Saint-Louis…   16.2    -15.3
#> # ℹ 23 more rows
#> # ℹ 2 more variables: nautical_miles <int>, status <chr>