Flights of U.S. domestic airlines February 2011

A dataset containing map coordinates from the departures and arrivals at U.S. airports in February of 2011.

Usage

flights

Format

A data frame with 178 rows and 9 columns:

start_lat

latitude of the departing airport

start_lon

longitude of the departing airport

end_lat

latitude at the arriving airport

end_lon

longitude at the arriving airport

airline

airline code

airport1

three letter abbreviation for departing airport

airport2

three letter abbreviation for arriving airport

cnt

number of flights

info

html label for origin and destination of flight

Source

'https://raw.githubusercontent.com/plotly/datasets/master/2011_february_aa_flight_paths.csv'

Details

This dataset is the illustrative dataset for the mapdeck package and was ported from the plotly GitHub account. To avoid the creation of a dependency and a possible broken link, the dataset was preserved here. The original source was provided under an MIT license.

Examples

if (requireNamespace("dplyr", quietly = TRUE)) {
  library(dplyr, warn.conflicts = FALSE, quietly = TRUE)
  data(flights)
  glimpse(flights)
}
#> Rows: 178
#> Columns: 9
#> $ start_lat <dbl> 32.89595, 41.97960, 32.89595, 18.43942, 32.89595, 25.79325, …
#> $ start_lon <dbl> -97.03720, -87.90446, -97.03720, -66.00183, -97.03720, -80.2…
#> $ end_lat   <dbl> 35.04022, 30.19453, 41.93887, 41.93887, 33.56294, 36.12448, …
#> $ end_lon   <dbl> -106.60919, -97.66987, -72.68323, -72.68323, -86.75355, -86.…
#> $ airline   <chr> "AA", "AA", "AA", "AA", "AA", "AA", "AA", "AA", "AA", "AA", …
#> $ airport1  <chr> "DFW", "ORD", "DFW", "SJU", "DFW", "MIA", "DFW", "MIA", "ORD…
#> $ airport2  <chr> "ABQ", "AUS", "BDL", "BDL", "BHM", "BNA", "BOS", "BOS", "BOS…
#> $ cnt       <int> 444, 166, 162, 56, 168, 56, 422, 392, 430, 56, 44, 112, 56, …
#> $ info      <chr> "<b>DFW - ABQ</b>", "<b>ORD - AUS</b>", "<b>DFW - BDL</b>", …