Homicides in San Francisco 2018 - 2023

Homicides in San Francisco, California, 2018 - 2023.

Usage

homicides

Format

a dataframe with 135 rows and 7 variables:

incident_datetime

date and time of the incident

incident_id

unique identifier for the incident

incident_code

incident code

incident_category

incident category

incident_subcategory

incident subcategory

lat

latitude

lon

longitude

Source

https://data.sfgov.org/Public-Safety/Police-Department-Incident-Reports-2018-to-Present/wg3w-h783/data_preview Published under a Open Data Commons Public Domain Dedication and License v1.0 (PDDL).

Details

For city-level crime data, municipalities, particularly large ones, often provide the data via an open data portal. Here, the data were discovered in the San Francisco Open Data Portal. Most crime data is reported on the Uniform Crime Reports and submitted to the Federal Bureau of Investigation. You can investigate and learn more about crime in the United States via the FBI's Crime Data Explorer. Murder is usually grouped with non-negligent manslaughter in counting homicides. For that reason, the "incident_subcategory" was included.

References

San Francisco Police Department, Police Department Incident Reports: 2018 to Present (2024) last accessed October 22, 2024.

Examples

homicides
#> # A tibble: 135 × 7
#>    incident_datetime   incident_id incident_code incident_category
#>    <dttm>                    <dbl> <chr>         <chr>            
#>  1 2023-12-28 05:38:00     1351016 01003         Homicide         
#>  2 2023-12-26 06:34:00     1354814 01160         Homicide         
#>  3 2023-12-26 06:34:00     1354978 01160         Homicide         
#>  4 2023-12-15 11:41:00     1347453 01001         Homicide         
#>  5 2023-12-13 00:00:00     1346676 01001         Homicide         
#>  6 2023-12-13 00:00:00     1346899 01001         Homicide         
#>  7 2023-12-13 00:00:00     1346906 01001         Homicide         
#>  8 2023-12-11 14:50:00     1350508 01160         Homicide         
#>  9 2023-12-06 01:46:00     1344886 01160         Homicide         
#> 10 2023-11-27 00:20:00     1340297 01001         Homicide         
#> # ℹ 125 more rows
#> # ℹ 3 more variables: incident_subcategory <chr>, lat <dbl>, lon <dbl>