U.S. life expectancy by county 2019

U.S. life expectancy by county for all races and sexes at less than one year in 2019.

Usage

life_exp

Format

A data frame with 3,127 rows and 9 variables:

year

year

state_fips

two digit state fips code

state_name

state name

fips

five digit county fips code

race_name

race name

sex_name

sex name

location_name

location name given as county and state

age_name

life expectancy at less than one year

life_exp

life expectancy in years

Source

https://ghdx.healthdata.org/record/ihme-data/united-states-causes-death-life-expectancy-by-county-race-ethnicity-2000-2019. No license was referenced.

Details

The data is from the Institute for Health Metrics and Evaluation (IHME) and is based on the Global Burden of Disease (GBD) study. The data is for the year 2019 and is at the county level.

References

Institute for Health Metrics and Evaluation (IHME). United States Mortality Rates by Causes of Death and Life Expectancy by County, Race, and Ethnicity 2000-2019. Seattle, United States of America: Institute for Health Metrics and Evaluation (IHME), 2023. doi.org/10.6069/3WQ2-TG23

See also

Examples

life_exp
#> # A tibble: 3,127 × 9
#>    year  state_fips state_name fips  race_name sex_name location_name   age_name
#>    <chr> <chr>      <chr>      <chr> <chr>     <chr>    <chr>           <chr>   
#>  1 2019  01         Alabama    01001 Total     Both     Autauga County… <1 year 
#>  2 2019  01         Alabama    01003 Total     Both     Baldwin County… <1 year 
#>  3 2019  01         Alabama    01005 Total     Both     Barbour County… <1 year 
#>  4 2019  01         Alabama    01007 Total     Both     Bibb County (A… <1 year 
#>  5 2019  01         Alabama    01009 Total     Both     Blount County … <1 year 
#>  6 2019  01         Alabama    01011 Total     Both     Bullock County… <1 year 
#>  7 2019  01         Alabama    01013 Total     Both     Butler County … <1 year 
#>  8 2019  01         Alabama    01015 Total     Both     Calhoun County… <1 year 
#>  9 2019  01         Alabama    01017 Total     Both     Chambers Count… <1 year 
#> 10 2019  01         Alabama    01019 Total     Both     Cherokee Count… <1 year 
#> # ℹ 3,117 more rows
#> # ℹ 1 more variable: life_exp <dbl>