Temperature data along Napolean's march to and from Moscow 1812

Temperature played a key role in the French Army's casualties as the cold Russian winters took a toll.

Usage

minard_temps

Format

A data frame of 8 observations of four variables:

date

date in the format of "YYYY-MM-DD"

temp

temperature

days

number of days on the retreat march

lon

longitude

Source

https://vincentarelbundock.github.io/Rdatasets/doc/HistData/Minard.troops.html; https://www.cs.uic.edu/~wilkinson/TheGrammarOfGraphics/minard.txt

Details

Original documentation of this dataset directs the date to be set to the year 1815. Historical context suggests that the data is from 1812 and the year was set as such.

References

Minard, Charles Joseph. 1869. "Tableaux graphiques et cartes figuratives." Paris: Hachette.

See also

Examples

minard_temps
#> # A tibble: 8 × 4
#>   date        temp  days   lon
#>   <date>     <dbl> <dbl> <dbl>
#> 1 1815-10-18     0     6  37.6
#> 2 1815-10-24     0     6  36  
#> 3 1815-11-09    -9    16  33.2
#> 4 1815-11-14   -21     5  32  
#> 5 1815-11-28   -20     4  28.5
#> 6 1815-12-01   -24     3  27.2
#> 7 1815-12-06   -30     5  26.7
#> 8 1815-12-07   -26     1  25.3