French Army survivors after key battles to and from Moscow 1812

Charles Joseph Minard's famous graphic of Napolean's march to Moscow shows the dwindling number of French Army troops.

Usage

minard_troops

Format

A data frame of 51 rows and five variables:

survivors

numeric value of French Army survivors

direction

a factor with levels A ("Advance") and R ("Retreat")

group

group

lon

longitude

lat

latitude

Source

https://vincentarelbundock.github.io/Rdatasets/doc/HistData/Minard.troops.html; https://www.cs.uic.edu/~wilkinson/TheGrammarOfGraphics/minard.txt

Details

Napoleon's invasion and subsequent retreat from Russia is notable for the significant loss of life. While exact numbers are elusive, historians estimate that Napoleon began the campaign with 450,000 troops and ended with 120,000. Napoleon's eastward progress ended in Moscow in November of 1812. Inadequate supplies required the French Army to retreat but only after the onset of winter like conditions. Frigid temperatures and disease decimated the remaining army. Charles Joseph Minard, an early pioneer in visualization, captured the French Army's losses in his famous graphic that has been called "the best statistical graphic ever drawn" by Edward Tufte.

References

Friendly, M. (2002). Visions and Re-visions of Charles Joseph Minard, Journal of Educational and Behavioral Statistics, 27, No. 1, 31-51.

Friendly, M. (2003). Re-Visions of Minard. http://datavis.ca/gallery/re-minard.html

Minard, Charles Joseph. 1869. "Tableaux graphiques et cartes figuratives." Paris: Hachette.

Arthur H. Robinson (1967), "The Thematic Maps of Charles Joseph Minard", Imago Mundi, Vol. 21, (1967), pp. 95–108

Edward Tufte (1983), "The Visual Display of Quantitative Information", Graphics Press, Cheshire, Connecticut.

See also

Examples

minard_troops
#> # A tibble: 51 × 5
#>    survivors direction group   lon   lat
#>        <dbl> <chr>     <dbl> <dbl> <dbl>
#>  1    340000 A             1  24    54.9
#>  2    340000 A             1  24.5  55  
#>  3    340000 A             1  25.5  54.5
#>  4    320000 A             1  26    54.7
#>  5    300000 A             1  27    54.8
#>  6    280000 A             1  28    54.9
#>  7    240000 A             1  28.5  55  
#>  8    210000 A             1  29    55.1
#>  9    180000 A             1  30    55.2
#> 10    175000 A             1  30.3  55.3
#> # ℹ 41 more rows