White shark tracking data near Guadulope Island 2015 - 2019

A dataset containing coordinates for 10 great white sharks near Guadulope Island, Mexico, including depth.

Usage

white_shark

Format

A data frame with 322 rows and 10 variables:

shark

shark ID

date_hour

date and hour

obs

observations

n

nunmber of observations within original time increment

mean_depth

mean depth in meters

lat

latitude

lng

longitude

mean_shore_distance

mean distance from shore in meters

length

length of shark in meters

sex

sex of shark

Source

https://www.seanoe.org/data/00722/83385/

Details

Guadulope Island, off the coast of Mexico, is renown for its ecotourism around great white shark cage diving. The data was included because of the depth variable. The original dataset was +28,000k observations with the tracking device giving a reading nearly 20 times a minute. Here, the data were summarized to the hour.

References

Santana-Morales, O., Hoyos-Padilla, E. M., Medellín-Ortíz, A., Sepulveda, C., Beas-Luna, R., Aquino-Baleytó, M., ... & Castillo-Géniz, J. L. (2021). How much is too much? A carrying capacity study of white shark cage diving in Guadalupe Island, Mexico. Marine Policy, 131, 104588.

Examples

white_shark
#> # A tibble: 322 × 10
#>    shark date_hour   obs     n mean_depth   lat   lng mean_shore_distance length
#>    <fct> <chr>     <int> <int>      <dbl> <dbl> <dbl>               <dbl>  <dbl>
#>  1 WS1   2015-09-…     1   479      16.0   29.1 -118.                282.      4
#>  2 WS1   2015-09-…     2   374       1.71  29.1 -118.                454.      4
#>  3 WS1   2015-09-…     3   680      36.0   29.1 -118.                509.      4
#>  4 WS1   2015-09-…     4  1073      22.4   29.1 -118.                921.      4
#>  5 WS1   2015-09-…     5  1027       8.96  29.1 -118.                396.      4
#>  6 WS1   2015-09-…     6   971      17.5   29.1 -118.                717.      4
#>  7 WS1   2015-09-…     7   532      15.0   29.1 -118.                345.      4
#>  8 WS1   2015-09-…     8   133       1.93  29.1 -118.                857.      4
#>  9 WS1   2015-09-…     9    92       4.84  29.1 -118.                330.      4
#> 10 WS1   2015-09-…    10   205       5.85  29.1 -118.                279.      4
#> # ℹ 312 more rows
#> # ℹ 1 more variable: sex <chr>